requirements.txt
loguru
matplotlib
mlflow
numpy
omegaconf
pandas
Pillow
PyYAML
scikit_learn
timm
tqdm
